refactor(client): move ai agent to services folder
